Fix extension() to return the "." as in boost::filesystem
extension() returned only the substring after the final dot, making impossible to distinguish "foo." from "foo".
Changed to return the substring including the final dot, as in boost::filesystem, so that path("file.h").extension() returns ".h".

Return an empty string for dot-files - files or directories starting with a dot.